adds dummy nose and deactivates mesh renderer
I deactivated the mesh renderer component of the route waypoints. They are still visible when selected in the editor, but cannot be seen in game mode. Maybe there is a better way to make them properly visible during editing.

Also, I added a cube the front of the fish to see the face orientation. When we add a model later on to the fish, we want them swim with their face towards the next waypoint. Currently the always look in the same direction.