adds game states and transitions via main menu & game over screen
1. Starting and quitting the game via main menu possible.
From GameOver Screen going to the MainMenu and starting a new game there is possible.
A new GameState class is added. It uses an enum and a static state variable to only allow certain transitions from one game state to another. This class resembles a state design pattern or a finite state machine.
2. Also a directional light is activated (was already in the scene). Somehow the unity editor is adding some kind of debugging light automatically when starting a scene directly, therefore an additional light was needed.

Not implemented yet:
- Retrying from the game over screen
- Pausing the game via Esc key and a pause menu