playable directly from editor
The GameState class makes sure, that the internal state of the game is always aligned with the current scene. Usually the game is started via the main menu and the state instantiated accordingly. To play from the editor some special logic in `OnEnable()` and `OnSceneLoaded()` is applied. This was already implemented in code. Somehow one empty gameobject with the script was missing in the scene. This commit just adds the gameobject with script.